pyproject.toml
[tool.poetry]
name = "zinc"
version = "0.1.0"
description = "zrepl console"
authors = ["Jeff Hutchison <jeff@j3ff.io>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.0.2"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]
black = "^21.9b0"
pytest = "^6.2.5"

[tool.black]
line-length = 88
target-version = ['py39']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
